refactor: move code to internal folder as per convention
