+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"portal/config"
+	"portal/internal/config"
 )
 
 func main() {
